Update cache size data type to int64

+++ cluster/shardmgr.go
@@ -28,7 +28,7 @@ type ShardManagerConfig struct {
 	// Shard timeout in seconds
 	ShardTimeout int `yaml:"shardTimeout"`
 	// Cache size in bytes, set to -1 for unlimited, 0 for no shared caching
-	MaxCacheSize int32 `yaml:"maxCacheSize"`
+	MaxCacheSize int64 `yaml:"maxCacheSize"`
 }
 
 type ShardManager struct {
